Move login tests to their own file
